Run mypy on parts of the tests/ directory
Some of the test code already has type annotations, so might as well
verify them. Running mypy caught some mislabled types in the tests/lib/
directory.

The tests are a consumer of the pip functions and API. By running mypy
on tests, it helps ensure that tests are representative and that typing
is consistent with the expectations represented by tests.

As time goes on, the project can expand the mypy coverage of tests.

For the tests directory, untyped defs are allowed so as not to require
all tests functions be annotated.